+++ docs/conf.py
@@ -47,7 +47,6 @@ extensions = [
 
 autoclass_content = "both"
 autodoc_member_order = "groupwise"
-autodoc_type_aliases = {"JSON": "JSON"}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = []
@@ -76,5 +75,8 @@ python_use_unqualified_type_names = True
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
     "eth_account": ("https://eth-account.readthedocs.io/en/v0.13.0/", None),
+    "eth_typing": ("https://eth-typing.readthedocs.io/en/stable/", None),
     "ethereum_rpc": ("https://ethereum-rpc.readthedocs.io/en/v0.1.0", None),
+    "alysis": ("https://alysis.readthedocs.io/en/v0.6.0/", None),
+    "trio": ("https://trio.readthedocs.io/en/v0.31.0/", None),
 }