+++ pons/_abi_types.py
@@ -6,11 +6,14 @@ from abc import ABC, abstractmethod
 from collections.abc import Iterable, Mapping, Sequence
 from functools import cached_property
 from types import EllipsisType
-from typing import Any
+from typing import Any, cast
 
 import eth_abi
 from eth_abi.exceptions import DecodingError
 from ethereum_rpc import Address, keccak
+
+ABI_JSON = None | bool | int | float | str | Sequence["ABI_JSON"] | Mapping[str, "ABI_JSON"]
+"""Values serializable to JSON."""
 
 # Maximum bits in an `int` or `uint` type in Solidity.
 MAX_INTEGER_BITS = 256
@@ -147,6 +150,9 @@ class UInt(Type):
     def __eq__(self, other: object) -> bool:
         return isinstance(other, UInt) and self._bits == other._bits
 
+    def __hash__(self) -> int:
+        return hash((type(self), self._bits))
+
 
 class Int(Type):
     """Corresponds to the Solidity ``int<bits>`` type."""
@@ -183,6 +189,9 @@ class Int(Type):
     def __eq__(self, other: object) -> bool:
         return isinstance(other, Int) and self._bits == other._bits
 
+    def __hash__(self) -> int:
+        return hash((type(self), self._bits))
+
 
 class Bytes(Type):
     """Corresponds to the Solidity ``bytes<size>`` type."""
@@ -199,8 +208,7 @@ class Bytes(Type):
     def _check_val(self, val: Any) -> bytes:
         if not isinstance(val, bytes):
             raise TypeError(
-                f"`{self.canonical_form}` must correspond to a bytestring, "
-                f"got {type(val).__name__}"
+                f"`{self.canonical_form}` must correspond to a bytestring, got {type(val).__name__}"
             )
         if self._size is not None and len(val) != self._size:
             raise ValueError(f"Expected {self._size} bytes, got {len(val)}")
@@ -236,6 +244,9 @@ class Bytes(Type):
     def __eq__(self, other: object) -> bool:
         return isinstance(other, Bytes) and self._size == other._size
 
+    def __hash__(self) -> int:
+        return hash((type(self), self._size))
+
 
 class AddressType(Type):
     """
@@ -250,8 +261,7 @@ class AddressType(Type):
     def _normalize(self, val: Any) -> str:
         if not isinstance(val, Address):
             raise TypeError(
-                f"`address` must correspond to an `Address`-type value, "
-                f"got {type(val).__name__}"
+                f"`address` must correspond to an `Address`-type value, got {type(val).__name__}"
             )
         return val.checksum
 
@@ -262,6 +272,9 @@ class AddressType(Type):
 
     def __eq__(self, other: object) -> bool:
         return isinstance(other, AddressType)
+
+    def __hash__(self) -> int:
+        return hash(type(self))
 
 
 class String(Type):
@@ -299,6 +312,9 @@ class String(Type):
     def __eq__(self, other: object) -> bool:
         return isinstance(other, String)
 
+    def __hash__(self) -> int:
+        return hash(type(self))
+
 
 class Bool(Type):
     """Corresponds to the Solidity ``bool`` type."""
@@ -322,6 +338,9 @@ class Bool(Type):
 
     def __eq__(self, other: object) -> bool:
         return isinstance(other, Bool)
+
+    def __hash__(self) -> int:
+        return hash(type(self))
 
 
 class Array(Type):
@@ -365,6 +384,9 @@ class Array(Type):
             and self._element_type == other._element_type
             and self._size == other._size
         )
+
+    def __hash__(self) -> int:
+        return hash((type(self), self._element_type, self._size))
 
 
 class Struct(Type):
@@ -426,6 +448,9 @@ class Struct(Type):
             and list(self._fields) == list(other._fields)
         )
 
+    def __hash__(self) -> int:
+        return hash((type(self), tuple(self._fields.items())))
+
 
 _UINT_RE = re.compile(r"uint(\d+)")
 _INT_RE = re.compile(r"int(\d+)")
@@ -451,8 +476,11 @@ def type_from_abi_string(abi_string: str) -> Type:
     raise ValueError(f"Unknown type: {abi_string}")
 
 
-def dispatch_type(abi_entry: Mapping[str, Any]) -> Type:
-    type_str = abi_entry["type"]
+def dispatch_type(abi_entry: ABI_JSON) -> Type:
+    # TODO (#83): use proper validation
+    abi_entry_typed = cast("Mapping[str, Any]", abi_entry)
+
+    type_str = abi_entry_typed["type"]
     match = re.match(r"^([\w\d\[\]]*?)(\[(\d+)?\])?$", type_str)
     if not match:
         raise ValueError(f"Incorrect type format: {type_str}")
@@ -464,26 +492,29 @@ def dispatch_type(abi_entry: Mapping[str, Any]) -> Type:
         array_size = int(array_size)
 
     if is_array:
-        element_entry = dict(abi_entry)
+        element_entry = dict(abi_entry_typed)
         element_entry["type"] = element_type_name
         element_type = dispatch_type(element_entry)
         return Array(element_type, array_size)
 
     if element_type_name == "tuple":
         fields = {}
-        for component in abi_entry["components"]:
+        for component in abi_entry_typed["components"]:
             fields[component["name"]] = dispatch_type(component)
         return Struct(fields)
 
     return type_from_abi_string(element_type_name)
 
 
-def dispatch_types(abi_entry: Iterable[dict[str, Any]]) -> list[Type] | dict[str, Type]:
-    names = [entry["name"] for entry in abi_entry]
+def dispatch_types(abi_entry: ABI_JSON) -> list[Type] | dict[str, Type]:
+    # TODO (#83): use proper validation
+    abi_entry_typed = cast("Iterable[dict[str, Any]]", abi_entry)
+
+    names = [entry["name"] for entry in abi_entry_typed]
 
     # Unnamed arguments; treat as positional arguments
     if names and all(not name for name in names):
-        return [dispatch_type(entry) for entry in abi_entry]
+        return [dispatch_type(entry) for entry in abi_entry_typed]
 
     if any(not name for name in names):
         raise ValueError("Arguments must be either all named or all unnamed")
@@ -491,7 +522,7 @@ def dispatch_types(abi_entry: Iterable[dict[str, Any]]) -> list[Type] | dict[str
     # Since we are returning a dictionary, need to be sure we don't silently merge entries
     if len(names) != len(set(names)):
         raise ValueError("All ABI entries must have distinct names")
-    return {entry["name"]: dispatch_type(entry) for entry in abi_entry}
+    return {entry["name"]: dispatch_type(entry) for entry in abi_entry_typed}
 
 
 def encode_args(*types_and_args: tuple[Type, Any]) -> bytes: