+++ docs/changelog.rst
@@ -2,6 +2,27 @@ Changelog
 ---------
 
 
+0.9.0 (in development)
+~~~~~~~~~~~~~~~~~~~~~~
+
+Changed
+~~~~~~~
+
+- ``JSON`` removed from the public API, instead we have a more specific ``ABI_JSON``. (PR_82_)
+- Renamed ``id_`` fields of ``BlockFilter``, ``PendingTransactionFilter``, and ``LogFilter`` to just ``id``. (PR_82_)
+- Split out ``http-provider-server`` feature from ``local-provider``. (PR_82_)
+
+
+Added
+^^^^^
+
+- Hash methods for ABI types. (PR_81_)
+
+
+.. _PR_81: https://github.com/fjarri-eth/pons/pull/81
+.. _PR_82: https://github.com/fjarri-eth/pons/pull/82
+
+
 0.8.1 (2024-11-12)
 ~~~~~~~~~~~~~~~~~~